added dao and model objects

+++ src/main/java/com/example/ecommerce/dao/ProductDao.java
@@ -3,5 +3,5 @@ package com.example.ecommerce.dao;
 import com.example.ecommerce.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDao extends JpaRepository<Product, Long> {
+public interface ProductDao extends JpaRepository<Product, Integer> {
 }
